requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
streamlit==1.41.1
requests==2.32.3
pydantic==2.10.4
pydantic-settings==2.7.1
python-dotenv==1.0.1
langchain==0.2.11
langchain-groq==0.1.9
langchain-chroma==0.1.2
chromadb==0.5.23
sentence-transformers==3.0.1
mem0ai==0.1.61
pytest==8.3.4
httpx==0.28.1
